Force actions in check mode
Ansible modules command and shell do not support check mode. That means
that these tasks are skipped with --check enabled. If you want to use an output
of the task later in the playbook it won't work.

This forces the task to run even with '--check' - it executes the task just like in normal mode.

See:
docs.ansible.com/ansible/latest/user_guide/playbooks_checkmode.html

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>